requirements.txt
chromadb==0.5.5
fastapi==0.115.0
psycopg2-binary==2.9.9
pydantic==2.8.2
uvicorn==0.30.6
voyageai==0.2.3
